services/requirements.txt
# Cloud Run Services Requirements
# Common dependencies for all ingestion services

# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Google Cloud clients
google-cloud-storage==2.10.0
google-cloud-discoveryengine==0.11.9
google-cloud-aiplatform==1.38.0
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-vision==3.4.5

# Document processing
PyMuPDF==1.23.8

# Additional utilities
python-multipart==0.0.6
